requirements.txt
absl-py==0.9.0
adal==1.2.2
attrs==19.3.0
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.8
importlab==0.5.1
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
networkx==2.4
ninja==1.9.0.post1
oauth2client==4.1.3
Office365-REST-Python-Client==2.1.5
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.7
Pygments==2.5.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytype==2020.2.6
pytz==2019.3
PyYAML==5.3
requests==2.22.0
rsa==4.0
six==1.14.0
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
